# api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenKin Task Live-Resource Contract
  version: 0.1.0
  description: |
    Authoritative contract for the beta Web and iOS client surface, including
    task, approval, question, workspace, device-pairing, and live update APIs.
    Go handlers remain handwritten; TypeScript types are generated from this
    document. Relay requests add room and key query credentials.
servers:
  - url: /
paths:
...
  /api/pairing/sessions:
    post:
      operationId: createPairingSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairingSessionRequest"
      responses:
        "201":
          description: One-time pairing secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingSessionResponse"
...
components:
...
  schemas:
    PairingSessionRequest:
      type: object
      properties:
        label:
          type: string
    PairingSessionResponse:
      type: object
      required: [secret, expires_at]
      properties:
        secret:
          type: string
        expires_at:
          type: integer
...